feat: set analytics table as component
